==> Capstone Project 2/1. Data Wrangling.ipynb <==
# Data Wrangling

## Uploading Modules

import pandas as pd
import string
import re
import nltk
from nltk.tokenize import word_tokenize
from nltk.corpus import stopwords
#nltk.download('stopwords')
#nltk.download('punkt')
import emoji

## Loading the Data

active19 = pd.read_csv("active19.csv")
active20 = pd.read_csv("active20.csv")
lazy19 = pd.read_csv("lazy19.csv")
lazy20 = pd.read_csv("lazy20.csv")

After a text is obtained, we start with text normalization. Text normalization includes:
- converting all letters to lower or upper case
- converting numbers into words or removing numbers
- removing punctuations, accent marks and other diacritics
- removing white spaces
- expanding abbreviations
- removing stop words, sparse terms, and particular words
- text canonicalization

# Convert text to lowercase
active19['clean_text']=active19.Text.apply(lambda x: x.lower())
active20['clean_text']=active20.Text.apply(lambda x: x.lower())
lazy19['clean_text']=lazy19.Text.apply(lambda x: x.lower())
lazy20['clean_text']=lazy20.Text.apply(lambda x: x.lower())

# Remove numbers
active19['clean_text']=active19['clean_text'].apply(lambda x: re.sub(r'\d+', '', x))
active20['clean_text']=active20['clean_text'].apply(lambda x: re.sub(r'\d+', '', x))
lazy19['clean_text']=lazy19['clean_text'].apply(lambda x: re.sub(r'\d+', '', x))
lazy20['clean_text']=lazy20['clean_text'].apply(lambda x: re.sub(r'\d+', '', x))

# Remove punctuation
translator = str.maketrans('', '', string.punctuation)
active19['clean_text']=active19['clean_text'].apply(lambda x: x.translate(translator))
active20['clean_text']=active20['clean_text'].apply(lambda x: x.translate(translator))
lazy19['clean_text']=lazy19['clean_text'].apply(lambda x: x.translate(translator))
lazy20['clean_text']=lazy20['clean_text'].apply(lambda x: x.translate(translator))

# Remove whitespaces
active19['clean_text']=active19['clean_text'].apply(lambda x: x.strip())
active20['clean_text']=active20['clean_text'].apply(lambda x: x.strip())
lazy19['clean_text']=lazy19['clean_text'].apply(lambda x: x.strip())
lazy20['clean_text']=lazy20['clean_text'].apply(lambda x: x.strip())

# set 'SW' as english stopwords from NLTK and count them
SW = set(stopwords.words('english'))
# remove 'not' from set SW
SW.remove('not')
SW.add('amp') # add 'amp' that shows up in tweets after '&' as in '&amp' from set SW
SW.add('»') #left after punctuation removal
SW.add('”') #left after punctuation removal
SW.add('“') #left after punctuation removal
SW.add('’ ') #left after punctuation removal
SW.add('im')
SW.add("I'm")
SW.add('gon na')
print('Number of english stopwords', len(SW))

# view english stopwords
print(SW)

def remove_stop_words(x):
    words = word_tokenize(x) # make a list of words
    useful_words = [w for w in words if w not in SW]    # remove stopwords
    return (' '.join(useful_words))

# remove stopwords:
active19['clean_text'] = active19['clean_text'].apply(lambda x: remove_stop_words(x))
active20['clean_text'] = active20['clean_text'].apply(lambda x: remove_stop_words(x))
lazy19['clean_text'] = lazy19['clean_text'].apply(lambda x: remove_stop_words(x))
lazy20['clean_text'] = lazy20['clean_text'].apply(lambda x: remove_stop_words(x))

# lemmatize
import nltk
nltk.download('wordnet')
from nltk.stem import WordNetLemmatizer
from nltk.tokenize import word_tokenize
lemmatizer=WordNetLemmatizer()
#lemmatizer.lemmatize(word)

active19['clean_text'] = active19['clean_text'].apply(lambda x: lemmatizer.lemmatize(x))
active20['clean_text'] = active20['clean_text'].apply(lambda x: lemmatizer.lemmatize(x))
lazy19['clean_text'] = lazy19['clean_text'].apply(lambda x: lemmatizer.lemmatize(x))
lazy20['clean_text'] = lazy20['clean_text'].apply(lambda x: lemmatizer.lemmatize(x))

# convert emoji to text:
active19['clean_text'] = active19['clean_text'].apply(lambda x: emoji.demojize(x, use_aliases=False, delimiters=('', '')))
active20['clean_text'] = active20['clean_text'].apply(lambda x: emoji.demojize(x, use_aliases=False, delimiters=('', '')))
lazy19['clean_text'] = lazy19['clean_text'].apply(lambda x: emoji.demojize(x, use_aliases=False, delimiters=('', '')))
lazy20['clean_text'] = lazy20['clean_text'].apply(lambda x: emoji.demojize(x, use_aliases=False, delimiters=('', '')))

# remove ' and gon na:
active19['clean_text'] = active19['clean_text'].str.replace(r'gon na', 'going')
active20['clean_text'] = active20['clean_text'].str.replace(r'gon na', 'going')
lazy19['clean_text'] = lazy19['clean_text'].str.replace(r'gon na', 'going')
lazy20['clean_text'] = lazy20['clean_text'].str.replace(r'gon na', 'going')

active19['clean_text'] = active19['clean_text'].str.replace(r'’ ', '')
active20['clean_text'] = active20['clean_text'].str.replace(r'’ ', '')
lazy19['clean_text'] = lazy19['clean_text'].str.replace(r'’ ', '')
lazy20['clean_text'] = lazy20['clean_text'].str.replace(r'’ ', '')

lazy20.head()

## Files for Future Use

#save dataframes as csv for future analysis
active20.reset_index().to_csv("active20-clean.csv")
active19.reset_index().to_csv("active19-clean.csv")
lazy20.reset_index().to_csv("lazy20-clean.csv")
lazy19.reset_index().to_csv("lazy19-clean.csv")
